More folder structure refactoring
