+++ OatMilk.Backend.Api/Data/Entities/Ability.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace OatMilk.Backend.Api.Data.Models.Entities
+namespace OatMilk.Backend.Api.Data.Entities
 {
     public class Ability
     {
